requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pyautogen>=0.2.0
pytest>=7.0.0
pytest-asyncio>=0.23.0 